requirements.txt
python-telegram-bot[webhooks]~=21.11
google-genai~=1.14
pydub~=0.25
